move read_obj to module readers
