+++ common/dal/Cargo.toml
@@ -28,8 +28,6 @@ rusoto_core = "0.47.0"
 rusoto_s3 = "0.47.0"
 rusoto_sts = "0.47.0"
 rusoto_credential = "0.47.0"
-serde = { version = "1.0.130", features = ["derive"] }
-serde_json = "1.0.72"
 
 [dev-dependencies]
 common-metrics= {path = "../metrics"}
